+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.common.core
 - Value.clear() added, PropertyValue.clear() renamed remove() in order to not clash with method from supertype.
 - Value.Builder, methods for adding listeners and consumers added.
+- ValueSet.Builder, methods for adding listeners and consumers added.
 ### is.codion.swing.common.ui
 - InputDialogBuilder.inputValid() and inputValidator() renamed valid() and validator() respectively.
 ### is.codion.swing.fraework.ui